requirements.txt
django
Pillow
pandas
openai
numpy
python-dotenv
requests
